% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_convert.R
\name{bd_wgs}
\alias{bd_wgs}
\title{bd09 to WGS84 coordinate system}
\usage{
bd_wgs(lon, lat)
}
\arguments{
\item{lon}{A numeric scalar or vector of longitude with bd09 as crs}

\item{lat}{A numeric scalar or vector of latitude with bd09 as crs}
}
\value{
A list contining X and Y numeric vectors
}
\description{
bd09 to WGS84 coordinate system
}
\examples{
data.frame(bd_wgs(105, 33))
}
