% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_convert.R
\name{bd_wgs_accu}
\alias{bd_wgs_accu}
\title{bd09 to WGS84 coordinate system with more accurate algorithm}
\usage{
bd_wgs_accu(lon, lat)
}
\arguments{
\item{lon}{A numeric scalar or vector longitude with gcj02 as crs}

\item{lat}{A numeric scalar or vector latitude with gcj02 as crs}
}
\value{
A list contining X and Y numeric vectors
}
\description{
bd09 to WGS84 coordinate system with more accurate algorithm
}
\examples{
data.frame(bd_wgs_accu(lon = 111:112,lat = 34:35))
}
