% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bd_wgs_accu_}
\alias{bd_wgs_accu_}
\title{More accurately convert bd09 to WGS84 coordinate system}
\usage{
bd_wgs_accu_(lon, lat)
}
\arguments{
\item{lon}{input longitude with bd09 as crs}

\item{lat}{input latitude with bd09 as crs}
}
\value{
A numeric vector with 2 elements contining X and Y numeric vectors
}
\description{
More accurately convert bd09 to WGS84 coordinate system
}
