% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_convert.R
\name{gcj_wgs_accu}
\alias{gcj_wgs_accu}
\title{gcj02 to WGS84 coordinate system with more accurate algorithm}
\usage{
gcj_wgs_accu(lon, lat)
}
\arguments{
\item{lon}{A numeric scalar or vector of longitude with gcj02 as crs}

\item{lat}{A numeric scalar or vector of latitude with gcj02 as crs}
}
\value{
A list contining X and Y numeric vectors
}
\description{
gcj02 to WGS84 coordinate system with more accurate algorithm
}
\examples{
data.frame(gcj_wgs_accu(111:112,34:35))
}
