% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gcj_wgs_accu_}
\alias{gcj_wgs_accu_}
\title{More accurately convert gcj02 to WGS84 coordinate system}
\usage{
gcj_wgs_accu_(lon, lat)
}
\arguments{
\item{lon}{input longitude with gcj02 as crs}

\item{lat}{input latitude with gcj02 as crs}
}
\value{
A numeric vector with 2 elements contining X and Y numeric vectors
}
\description{
More accurately convert gcj02 to WGS84 coordinate system
}
