% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_convert.R
\name{wgs_bd}
\alias{wgs_bd}
\title{WGS84 to bd09 coordinate system}
\usage{
wgs_bd(lon, lat)
}
\arguments{
\item{lon}{A numeric scalar or vector of longitude with WGS84 as crs}

\item{lat}{A numeric scalar or vector of latitude with WGS84 as crs}
}
\value{
list contining X and Y numeric vectors
}
\description{
WGS84 to bd09 coordinate system
}
\examples{
data.frame(wgs_bd(105, 33))
}
