% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_convert.R
\name{wgs_gcj}
\alias{wgs_gcj}
\title{WGS84 to gcj02 coordinate system}
\usage{
wgs_gcj(lon, lat)
}
\arguments{
\item{lon}{A numeric scalar or vector of longitude with WGS84 as crs}

\item{lat}{A numeric scalar or vector of latitude with WGS84 as crs}
}
\value{
list contining X and Y numeric vectors
}
\description{
WGS84 to gcj02 coordinate system
}
\examples{
data.frame(wgs_gcj(105, 33))
}
